vlc inital volume settings
Add initial volume settings for vlc player.
Sets the initial volume to a value determined in the mycroft.conf
or sets it to a sensible default (either initial volume or 50%)
independent of system volume.
Also checks wheter the volume of the player is lower than low_volume and
pauses the player if necesarry.

==== Fixed Issues ====
NONE

====  Tech Notes ====
VLC (and other media players) use the system volume when called.
This can be inapproiate when the media is (way) louder than mycroft,
leading to a difference in volume between speech and media
The typical user case is:
1) mycroft volume is appropiate
2) Start of a media (eg radio)
3) music is too load
4) Decrease the (overall) volume
5) When barking in, mycroft cannot be understood, because it's too low.

As of now, there is no way to communicate a specific volume
for a specific skill or media.
This is a workaround, when the local backend is
normally used by mycroft, but media is played by another backend(eg vlc)

====  Documentation Notes ====
adds "volume" config key in mycroft.conf under "vlc"
volume is an integer between 0 and 100, where 0 is mute and 100 is full
system volume.

Example:

"Audio": {
  "backends": {
    "vlc": {
       "volume": "60"
    }
  }
}

==== Localization Notes ====
NONE

==== Environment Notes ====
NONE

==== Protocol Notes ====
NONE